fix the blur shader (how the fuck did i not notice this??)

+++ renderer/src/main/java/me/x150/renderer/util/RenderUtils.java
@@ -258,6 +258,6 @@ public class RenderUtils {
 	 * @return The GUI scale of the current window
 	 */
 	public static int getGuiScale() {
-		return (int) MinecraftClient.getInstance().getWindow().getScaleFactor();
+		return MinecraftClient.getInstance().getWindow().getScaleFactor();
 	}
 }
